Change HTML Document to Not Self Contained
In order to make the HTML file size smaller, we changed the type from
self contained to not self contained. This is done in the hope that the
final HTML file renders on GitHub. This was not the case before, because
the file was deemed too big. If the file won't render properly either,
we may revert this change in a following commit.

Furthermore, we fixed the link in the REAMDE as it is containing a
whitespace and we need to encode whitespaces in links via `%20`.