Add index.Rmd
GitHub pages are now configured and running. This makes an index.html
necessary. An index.Rmd was created which simply stores the links to all
solution markdowns.

In a second step, the knitr hooks are adapted such that the index.Rmd is
updated and rendered whenever a solution markdown is rendered.