Add Solution for Day 3
Some simplifications have been incorporated into the RMarkdown file such
that copying will be less pain. For example a parameter has been added,
which holds the day number and one has to change this numebr only once
and links should then be refelcting this day.

Some trials with git hooks seem promising as a replacment for the
knit:render hack. Will investigate more, to see when an automated script
updating the indices should best be run.