Add Makfeile to Project for Docs Rendering
In this commit we added a Makefile which will handle the update of
REAMDE.md and index.Rmd (and consequently index.html). With this change
we can remove the logic from the knit hook in the Rmds. The final step
is to add a pre-commit hook, which runs make to make (pun intended) sure
that the docs are generated and part of the commit.