Add Solution for Day 5
Besides adding the solution for day 5, we also changed the Makefile and
the R script generating the documentation to
* find Pandoc
* use the right locale

With these settings, one can simply define a pre-commit hook calling
make and adding the docs to the index. DO not forget to add the shebank
to the beginning of the hook script.