Add Solution for Day 6
Besides the solution for day 6 itself. we also fixed an issue with the
auto doc script, where libs were not found. This happened b/c the
`--vanilla` option prevented `RScript` from using any startup scripts,
where the place of the additional libraries are defined. We solved the
issue by manually adding the library paths to `.libPaths`.