Add Solution Stub for Day 12
We use the `igraph` package for representing the graph. The algo
idea is to use a recursive function, which visits all neighbors.
In a certain path, we will visit a node again only if it is all
uppercase ("a big cave"). Thus, we will assign an attribute with
the visitability of a node and use that in our recursive algo.