Add Solution for Day 12
We used package `igraph`for this puzzle, but it proved to be slow. Thus,
a fallback option using plain R data structures was added (which brought
some significant speed gain). The speed overall, especiayll for teh
second puzzle is not great though, whihc hints that the used
backtracking algo is maybe not the bets solution.

We needed to adapt the build script as well to allow for paths at
differnt locations. This is a dirty hack and should better be solved
with environment variables.