Add Solution for Day 14
This puzzle wasa tough one, as I got on a foold's errand. Eventually, I
looke dup for some inspiration and found my major mistake in still
trying to brute-force the solution (yet more elegantly with a graph). In
the end I still used the graph solution, but instead of backtracing each
edge - which is simply too big a task - I used the idea of the solutions
I saw. We do not have to backtrace each edge, but we can simply keep an
updated count of the last visited nodes. All we have to do then is to
visit all the 2 neighbors of the nodes lasted visited which is a much
smaller task of course.