Add  Solution for Day 16
The puzzle was not overly clear where the padding happens, so it took me
a while to realize that the 0 padding can only occur at the end of the
last packet, Once this was clarified, the parser worked rather smoothly
and the second part was very easy due to R capabilities of `do.call` and
`...`. A neat trick was to use `quote` to store the `while` loop
condition. In this way, we did not repliacate the body of the while
loop though the differnt packets use different end condiitons (total
size vs. nr of packets.

The output is classed, b/c in the development version we defined some
helpers for printing.