Fix Build Script to Avoid Year Change
The build scripts were written with the idea in mind that all puzzles
will be solved only in the year when they are published, but should
allow for new puzzle solution to be added in the next years. However,
due to the commitments, I could not solve all the puzzles on the days
when they came out, but as I still want to finish them, I needed to
adapt the scripts to allow for manually overriding the defult to use the
actual year.