Add Solution fod Day 18
This was a tough nut to crack. There was a bug in my implemtation, which
did not show on most of the use cases. In principle hte funciton
identifxing the next node was faulty (it returned right nodes before
left nodes if they were at a lower depth). This bug was (at least for
me) imposisble to spot by pure eye, so I got a working solution in
python from th net and compared the steps to identify the culprit.

After that the solution for a proper funciton took me again some time,
but ultimately I managed to get to teh correct solution. The selected R6
framework and my implementation is however, awefully slow. So the second
task took a couple of hours to just reduce 10000 numbers. For the sake
of learning I might add an CPP solution.