Add C++ Solution for Puzzle Day 18
A C++ solution was finally added to the code base, which mimics the R
solution but which is faster by magnitudes (instantaneous results versus
running times for hours). The main reason for that is that in the R
solution which relies on R6 `environments` are used (as backbone of R6).
The copy-on-modify and call-by-value semantics of R introduce a heavy
overload for the selected recursive data structure.

This massive bottleneck could be easily resolved via a lean class in
C++, where we do not have to copy data structures as heavily as in R.

I learned a lot about smart pointers (std::shared_ptr) and how to break
cycles via std::weak_ptr (cycles were needed, becasue any child element
needs also to point to its parent to allow for proper traversal of the
structure).

We also had to inherite from std::enable_shared_from_this to properly
return a pointer from this.

Eventually we also had to "patch" the selected Rmarkdown format, as in
the standard JS there was no support for C++ files. This has been
remedied by firts udnerstanding which JS is responsible for syntax
highlighting (`Prism.js`) and then sourcing the language extensions. As
the cpp syntax highlihter depends also on the "c" syntax highlighter
both needed to be sourced.