Add Solution for Task 19
The main idea of this solution was to find matching beacons via their
distance. Once common beacons were identified, the proper rotation and
the offset had to be determined. These rotations and offsets had to be
recursively applied to base everything on the same basis. Once this was
done, a count of unique beacons was all left to do.

The second part of the puzzle was easy, as we already aligned the coords
in the first part. So all we had to do was to calculate all pairwise
distances.