Add Solution for Day 22
My original solution was working (despite the need to change int to
numeric, whihc is ont done in the current patch), but it was awefully
slow, due to the overengineering via `S4` and growing lists.

I checked other soluitons on Reddit and got an idea of how to use my
same idea but much more efficiently, dropping the otherwise beautiful S4
solution.